requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.3
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
boto3==1.14.7
botocore==1.17.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-braces==1.14.0
django-filter==2.2.0
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docutils==0.15.2
drf-yasg==1.17.1
Faker==2.0.4
gunicorn==20.0.4
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
olefile==0.46
packaging==20.4
phonenumbers==8.12.1
Pillow==7.0.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
wincertstore==0.2
